requirements.txt
tensorflow
numpy
pandas
pillow
matplotlib
scikit-learn
openpyxl
PyPDF2